created Navbar component
