+++ src/components/Layout/styles.ts
@@ -1,26 +1,12 @@
-import { createStyles, makeStyles } from "@mui/styles"
-import { Theme } from "@mui/material/styles"
+import { makeStyles } from "@mui/styles"
 
-export const useStyles = makeStyles((theme: Theme) =>
-	createStyles({
-		navbar: {
-			backgroundColor: theme.palette.grey.A200,
-			"& a": {
-				color: "#fff",
-				marginLeft: 10,
-			},
-		},
-		mainContainer: {
-			paddingTop: "7rem",
-			paddingBottom: "7rem",
-			minHeight: "80vh",
-		},
-		footer: {
-			textAlign: "center",
-		},
-		brand: {
-			fontWeight: "bold",
-			fontSize: "1.5rem",
-		},
-	})
-)
+export const useStyles = makeStyles({
+	mainContainer: {
+		paddingTop: "7rem",
+		paddingBottom: "7rem",
+		minHeight: "80vh",
+	},
+	footer: {
+		textAlign: "center",
+	},
+})
